third_party/WebKit/LayoutTests/external/wpt/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_3-expected.txt

This is a testharness.js-based test.
FAIL Removing iframe from document removes it from history assert_array_equals: property 2, expected 1 but got 2
Harness: the test ran to completion.
